# metal : use FA-vec kernel up to batch size 20